Synch with HEAD:
MFC revision 1.167
date: 2005/08/08 21:06:42;  author: csjp;  state: Exp;  lines: +2 -0
Drop in a WITNESS_WARN into SYSCTL_IN to make sure that we are
not holding any non-sleep-able-locks locks when copyin is called.
This gets executed un-conditionally since we have no function
to wire the buffer in this direction.

MFC revision 1.166
date: 2005/08/08 18:54:35;  author: csjp;  state: Exp;  lines: +6 -1
Check to see if we wired the user-supplied buffers in SYSCTL_OUT, if
the buffer has not been wired and we are holding any non-sleep-able locks,
drop a witness warning. If the buffer has not been wired, it is possible
that the writing of the data can sleep, especially if the page is not in
memory. This can result in a number of different locking issues, including
dead locks.

Approved by:	re (kensmith)


git-svn-id: https://svn.freebsd.org/base/stable/6@149269 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f