MFC:	usr.sbin/jls/jls.c	1.4,1.5
Not sure why, but SYSCTL_OUT() can sometimes keep returning ENOMEM
in sysctl_jail_list(). Because of this, jls(8) could enter into
an endless loop. The strange thing is, that we can call jls(8) while
the other one is in loop and it will succeed - SYSCTL_OUT() will
not return ENOMEM there.

Approved by:	re (kensmith)


git-svn-id: https://svn.freebsd.org/base/stable/6@149277 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f