MFC:	sys/opencrypto/cryptodev.c	1.26
Fix bogus check. It was possible to panic the kernel by giving 0 length.
This is actually a local DoS, as every user can use /dev/crypto if there
is crypto hardware in the system and cryptodev.ko is loaded (or compiled
into the kernel).

Reported by:    Mike Tancsa <mike@sentex.net>
Security:	Local DoS on systems with crypto HW and with /dev/crypto available


git-svn-id: https://svn.freebsd.org/base/stable/5@149280 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f