Add some rationale about when to bump and not bump the config version. Clarify that it is not like the shlib versions, and not like param.h's __FreeBSD_version/osreldate either.
When config(8) was actively changing a while back, the interface between
config and the build system (eg: /sys/conf/files.* and Makefile.*) was
changing rapidly.  configvers is a version number of that interface.

User specified config files do not have a version number.  The decision
about whether a user supplied config file is syntactically valid or not
belongs to the parser and sanity checks, not an arbitary number.


git-svn-id: https://svn.freebsd.org/base/head@152162 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f