Fix a bug that caused some /dev entries to continue to exist after the underlying drive had been hot-unplugged from the system.  Here is a specific example.  Filesystem code had opened /dev/da1s1e. Subsequently, the drive was hot-unplugged.  This (correctly) caused all of the associated /dev/da1* entries to be deleted.  When the filesystem later realized that the drive was gone it closed the device, reducing the write-access counts to 0 on the geom providers for da1s1e, da1s1, and da1.  This caused geom to re-taste the providers, resulting in the devices being created again.  When the drive was hot-plugged back in, it resulted in duplicate /dev entries for da1s1e, da1s1, and da1.
This fix adds a new disk_gone() function which is called by CAM when a
drive goes away.  It orphans all of the providers associated with the
drive, setting an error condition of ENXIO in each one.  In addition,
we prevent a re-taste on last close for writing if an error condition
has been set in the provider.

Sponsored by:   Isilon Systems
Reviewed by:    phk
MFC after:      1 week


git-svn-id: https://svn.freebsd.org/base/head@152565 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f