"-o rw" is invalid and undocumented mount option that is only present for fstab(5) compatibility, and is otherwise ignored by mount(8) (not passed to mount_* programs, and not passed to nmount(2)).
"-u -o rw" worked with an old mount(8) with mount_ufs.c
because "-o rw" was stripped and simple "-u" caused an
update of UFS from read-only to read-write, due to
inability of mount(2) to track changes in options
(MNT_RDONLY is either set or not).

"-u" no longer causes the transition from RO to RW,
now that mount(8) was converted to use nmount(2), so
an explicit change to RW is required.  Keep up with
this change, and use "-uw" to mount root read-write.


git-svn-id: https://svn.freebsd.org/base/head@153029 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f