MFC: 1.26
In moduledir_readhints() cast the value returned by sizeof() to ssize_t
when checking whether it's greater than a struct stat st_size in order
to also catch the case when st_size is -1. Previously this check didn't
trigger on sparc64 when st_size is -1 (as it's the case for a file on
a bzipfs, TFTP server etc.), causing the content of the linker hints
file to be copied to memory referenced by a null-pointer.

PR:		91231


git-svn-id: https://svn.freebsd.org/base/stable/6@154716 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f