Reduce the field width by 1 for many numeric fields so that most fields cannot run into other fields or field descriptors.  If the value is too large to fit in the field width, then the output format is adjusted so that the value (usually) fits, but with fields running together externally this adjustment usually didn't help.  Mostly it doesn't matter to lose 1 digit of precision, but switching the output format is bad if it happens often or gives bogus units.  The loss of width is most serious for fields near "Csw" (which are also the ones which must often ran together) since these have a high variance and large values relative to the possible field widths so the switch occurs more often now, and for the memory size fields where the switch gives the bogus units kKB or MKB.
Now only the fields for r, p, d, s and w can run into each other.
These fields have width 3, and 3 cannot be reduced to 2 without losing
all precision when the value is between 100 and 999.

Trim "pdwake" to "pdwak" at think time now that it doesn't get clobbered
at runtime.  The manpage doesn't need to be changed for this because
it documents the clobbered descriptor, unlike for 4 other too-long
descriptors which only get clobbered if there are lots of interrupt
sources.

Trim "% busy" to "%busy" since most other descriptors for percentages
are spelled without the space and this change makes changing the widths
of the %busy fields unnecessary.


git-svn-id: https://svn.freebsd.org/base/head@155666 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f