Modify UNIX domain sockets to guarantee, and assume, that so_pcb is always defined for an in-use socket.  This allows us to eliminate countless tests of whether so_pcb is non-NULL, eliminating dozens of error cases.  For now, retain the call to sotryfree() in the uipc_abort() path, but this will eventually move to soabort().
These new assumptions should be largely correct, and will become more so
as the socket/pcb reference model is fixed.  Removing the notion that
so_pcb can be non-NULL is a critical step towards further fine-graining
of the UNIX domain socket locking, as the so_pcb reference no longer
needs to be protected using locks, instead it is a property of the socket
life cycle.


git-svn-id: https://svn.freebsd.org/base/head@156807 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f