Initialize the new members of struct ip_moptions as a defensive programming measure.
Note that whilst these members are not used by the ip_output()
path, we are passing an instance of struct ip_moptions here
which is declared on the stack (which could be considered a
bad thing).

ip_output() does not consume struct ip_moptions, but in case it
does in future, declare an in_multi vector on the stack too to
behave more like ip_findmoptions() does.


git-svn-id: https://svn.freebsd.org/base/head@158729 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f