Remove bzero/bcopy vestiges
Be cognizant as to whether we're running 2KLogin f/w in target mode and
do the appropriate loopid load based upon that.

Do a first cut (seems to work, at least for amd64) at 64 bit target
mode for fibre channel cards. We could probably also do it for SPI
cards, but that's not supported right now.


git-svn-id: https://svn.freebsd.org/base/head@158817 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f