Remove GEOM_GPT.
I encountered repeatable Unaligned access panics similar to:

fatal kernel trap:

    trap entry     = 0x4 (unaligned access fault)
    cpuid          = 0
    faulting va    = 0xfffffc0000fae1c2
    opcode         = 0x28
    register       = 0x1
    pc             = 0xfffffc00004d07dc
    ra             = 0xfffffc00004d0a1c
    sp             = 0xfffffe0031875bc0
    curthread      = 0xfffffc007da237a0
        pid = 2, comm = g_event

These occur when a floppy or a CD (the 6.1R install CD will do in most
cases) is in the drive during boot.  A floppy or CD inserted *after*
the system was booted did not give me this problem.  Interestingly
enough booting from the 6.1-R install CD sometimes caused the crash,
sometimes it did not.  When it did not panic an installation of 6.1-R
was possible without any incidents.  A subsequent boot from HD with
the CD still present sometimes(...) worked fine, sometimes paniced.

Taking GEOM_GPT out of the kernel config file allows booting without
incidents.  We do not need GPT on Alpha, so this is the easy way out.

Discussed with: phk


git-svn-id: https://svn.freebsd.org/base/stable/6@159007 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f