If NO_BIND_MTREE is set, we don't have anywhere to install the namedb files to, so don't try.
The PR indicates that NO_BIND_NAMED would also trigger this problem, but
I was unable to reproduce it.

This is going directly to RELENG_6 because testing shows that HEAD is not
affected by this issue, likely due to the renovation of the make knobs
in that branch.

PR:		misc/97188
Submitted by:	Peter Losher <Peter_Losher@isc.org>


git-svn-id: https://svn.freebsd.org/base/stable/6@159216 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f