MFC: 1.78, 1.79
Remove incorrect null_checkexp() routine.  This
will allow the NFS server to call vfs_stdcheckexp() on the exported nullfs
filesystem, not the underlying filesystem being nullfs mounted.
If the lower filesystem was not NFS exported, then the NFS exported
null filesystem would not work.
Modify MNT_UPDATE behavior for nullfs so that it does not
return EOPNOTSUPP if an "export" parameter was passed in.
This should allow nullfs mounts to be NFS exported.

PR:             kern/87906


git-svn-id: https://svn.freebsd.org/base/stable/6@159700 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f