Add some OpenSolaris compatibility definitions which are only visible if _SOLARIS_C_SOURCE is defined.
Add two function prototypes which are required to feed high-resolution
times to DTrace. DTrace requires it's own functions with the dtrace_
prefix so that it knows not to try and trace them. This is a rule that
code executed from the DTrace probe context must obey.

The two functions are only be compiled if the KDTRACE option is defined
to compile in kernel support for loading the DTrace modules.


git-svn-id: https://svn.freebsd.org/base/head@160987 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f