MFC rev. 1.71, 1.72:
Guess the number of arguments to a function somewhat better.
Now GCC likes to stick a "mov %eax, %FOO" instruction before
"addl $BAR, %esp" if the function just called returns an int,
which is a very common case in the kernel.

Return -1 from db_numargs() if number of args couldn't be guessed.
Use this later to indicate in backtrace output that args shown are
uncertain.


git-svn-id: https://svn.freebsd.org/base/stable/6@161320 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f