Create a 'ready' handler for each personality. The purpose of this handler is to able to be called after *all* attach and enable events are done.
We establish a SYSINIT hook to call this handler. The current usage for it
is to add scsi target resources *after* all enables are done. There seems
to be some dependencies between different halves of a dual-port with respect
to target mode.

Put in more meaningful event messages for some events- in particular
QUEUE FULL events so we can see what the queue depth was when the
IOC sent us this message.

MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@162133 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f