MFC: Fix the IPv4 multicast routing detach path. On interface detach whilst the MROUTER is running, the system would panic as described in the PR.
The fix in the PR is a good start, however, the other state associated
with the multicast forwarding cache has to be freed in order to avoid
leaking memory and other possible panics.

More care and attention is needed in this area.

PR:		kern/82882
Approved by:	re (rwatson)
Revs:		1.119 src/sys/netinet/ip_mroute.c


git-svn-id: https://svn.freebsd.org/base/stable/6@163072 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f