Fixed -mprofiler-epilogue.  The garbage collector apparently doesn't understand that non-local variables can never be collected, and when it collected the static variable for mexitcount_libfunc, gcc aborted on the next use of this variable.
This quick fix is to reinitialize the variable on every use and depend
on garbage collection recovering the small amount of memory wasted by
this, and not worry by the small amount of time wasted by this.  It
would be better to initialize the variable together with most of the
other libfuncs in optabs.c and depend on whatever magic is there to
prevent its collection, but we initialize it here to avoid taking at
least 2 more files off the vendor branch.


git-svn-id: https://svn.freebsd.org/base/head@163679 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f