Add a build option to support WITH_CDDL and WITHOUT_CDDL, defaulting to WITH_CDDL.
This option enables building code that is licensed under Sun's CDDL.
The DTrace code is licensed that way, so by default it will get built
unless the WITHOUT_CDDL option is used.

There is another build toggle, NO_CTF, which turns off execution of
ctfconvert and ctfmerge in sys.mk, but this can't be implemented as
WITH_/WITHOUT because bsd.own.mk isn't included in all Makefiles and
sys.mk is included automatically by make.


git-svn-id: https://svn.freebsd.org/base/head@163861 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f