MFC: 1.16
In the replacement text of the __bswapN_const() macros encapsulate the
argument in parentheses so these macros are safe to use and invocations
with an expression as the argument like __bswap32_const(42 << 23 | 13)
work as expected. Additionally, mask all the individually shifted bytes
as appropriate so the bytes which exceed the width of the respective
__bswapN_const() macro in invocations like __bswap16_const(0xdead600d)
are ignored like it's the case with the corresponding __bswapN_var()
function.

Approved by:	re (hrs)


git-svn-id: https://svn.freebsd.org/base/stable/6@163997 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f