Attempt to improve application portability by marking `struct ar_hdr' as `packed'.
The C standard leaves the alignment of individual members of a C
struct upto the implementation, so pedantically speaking portable
code cannot assume that the layout of a `struct ar_hdr' in memory
will match its layout in a file.  Using a __packed attribute
declaration forces file and memory layouts for this structure to
match.

Submitted by:	ru


git-svn-id: https://svn.freebsd.org/base/head@164244 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f