On RELENG_6, vn_open() will drop Giant aquired for calls into non-MP-safe filesystem if fdx == -1, assuming the in-kernel callers are not MP-safe. On HEAD, vn_open() after rev. 1.240 of sys/kern/vfs_vnops.c would keep the Giant locked when the MPSAFE flag is given.
The difference in the API caused the auditctl() syscall code to try to
unlock non-locked Giant for non-MP-safe filesystems, for instance, UFS
when QUOTA is compiled in.

This is direct commit to RELENG_6 to accomodate for RELENG_6' vn_open().

Reported by:	Steve Wills (steve at stevenwills com)
Reviewed by:	rwatson
Approved by:	re (bmah)
X-MFC to RELENG_6_2 after:	3 days


git-svn-id: https://svn.freebsd.org/base/stable/6@164387 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f