protect against multiple inclusion (this is useful when you start working with third party usb modules, where sometimes it is not easy to set the inclusion order so that there are no multiple inclusions, yet you want to compile with high WARNS levels).
I am not sure if there is a standard for having a leading and/or trailing _
in the macro name, the usb code seems to use both.

There are still several unprotected headers here so it might be useful
to do the same thing on other files as well as the need arises.

MFC After: 3 days


git-svn-id: https://svn.freebsd.org/base/head@165749 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f