MFC df.c revision 1.66
Based on The Open Group Base Specifications Issue 6 IEEE Std 1003.1, our
current implementation of df(1) is does not properly format the output under
certain conditions. Right now -kP and -Pk are not the same thing. Further,
when we set the BLOCKSIZE environment variable, we use "1k" instead of "1024",
making the header display incorrectly.


git-svn-id: https://svn.freebsd.org/base/stable/6@166114 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f