Cleaned up declaration and initialization of clock_lock.  It is only used by clock code, so don't export it to the world for machdep.c to initialize.  There is a minor problem initializing it before it is used, since although clock initialization is split up so that parts of it can be done early, the first part was never done early enough to actually work.  Split it up a bit more and do the first part as late as possible to document the necessary order.  The functions that implement the split are still bogusly exported.
Cleaned up initialization of the i8254 clock hardware using the new
split.  Actually initialize it early enough, and don't work around it
not being initialized in DELAY() when DELAY() is called early for
initialization of some console drivers.

This unfortunately moves a little more code before the early debugger
breakpoint so that it is harder to debug.  The ordering of console and
related initialization is delicate because we want to do as little as
possible before the breakpoint, but must initialize a console.


git-svn-id: https://svn.freebsd.org/base/head@166186 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f