MFC: 1.161      include the opt_diver.h to get the banner massage right.
This brings ip_fw2.c up to date with -current with the exception of:
 1.113, 114, 125, 126, 130, 138, 140, 152-156, 158-160
All of which I do not believe to be MFC candidates (with the
possible exception of 140 which has a risk factor I don't
want to put in RELENG_6). 138 is also a possibility but requires more work
in other files.
I need to do the userland ipfw too to add some new features, but, not today..


git-svn-id: https://svn.freebsd.org/base/stable/6@168110 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f