This trivial change should fix at least 3 similar bugs.  All of them are related to the `c' function's need to know if we are at the actual end of the address range.  (It must print the text not earlier than the whole pattern space was deleted.)  It appears the only sed function with this requirement.
There is `lastaddr' set by applies(), which is to notify the `c'
function, but it can't always help because it's false when we are
hitting the end of file early.  There is also a bug in applies()
due to which `lastaddr' isn't set to true on degenerate ranges such
as `$,$' or `N,$' if N appears the last line number.

Handling early EOF condition in applies() could look more logical,
but it would effectively revert sed to the unreasonable behaviour
rev. 1.26 of main.c fought against, as it would require lastline()
be called for each line within each address range.  So it's better
to call lastline() only if needed by the `c' function.

Together with this change to sed go regression tests for the bugs
fixed (c1-c3).  A basic test of `c' (c0) is also added as it helped
me to spot my own error.

Discussed with:		dds
Tested by:		the regression tests
MFC after:		1 week


git-svn-id: https://svn.freebsd.org/base/head@168211 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f