Implement the Marker Protocol. A marker frame is placed on the interface queue of each port and any further packets are blocked, when the all the marker frames have been returned to us from the remote network device then we can be sure that all interface queues are empty.
This is needed when a port is added or removed from the aggregation since it
will affect the hash based distribution, if the queues are not empty then a
packet from an existing connection may be placed on a different interface and
arrive out of order. This was previously achieved by suppressing transmission for
1 second, now that there is an active feedback this timeout as been increased
to 3 seconds and used as a fallback.


git-svn-id: https://svn.freebsd.org/base/head@169739 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f