MFC of the work committed on 2007-06-18 to improve the default named.conf. This brings us in line with BCP, and makes our installed configuration usable as a local resolver right out of the box.
This includes:
1.359 of etc/Makefile
1.5 of etc/namedb/Makefile
1.1 of etc/namedb/master/Makefile
all for the new files in etc/namedb/master,

1.23 and 1.24 of etc/namedb/named.conf to bring in the new
configuration options,

The new files in etc/namedb/master to support those options,

And deletion of the old *localhost*.rev files.

See the commit message to HEAD and the comments in the new
named.conf file for more details about the changes and the
reasons for them.


git-svn-id: https://svn.freebsd.org/base/stable/6@171356 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f