MFC rev. 1.710:
Use curthread instead of the FIRST_THREAD_IN_PROC for vnlru and syncer,
when applicable.

Aquire Giant slightly later for vnlru.

In the syncer, aquire the Giant only when a vnode belongs to the
non-MPsafe fs.

In both speedup_syncer() and syncer_shutdown(), remove the syncer thread from
the lbolt sleep queue after the syncer state is modified, not before.

Approved by:	re (kensmith)


git-svn-id: https://svn.freebsd.org/base/stable/7@174577 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f