Add support for disc3.  Depending on architecture the combination of kde3 and gnome2 is at least 670Mb which would mean if we stick with just disc1/disc2 the only packages we could provide on the iso's would be X11, kde3, and gnome2...
This is not an MFC, it was not committed to HEAD or RELENG_7.  It's
looking like we should plan on a bit more radical a change for how
packages get done for 8.X (current thought is cdrom sized things with
no packages at all, and one DVD sized thing with packages).  It's
almost certain this change will be done in RELENG_7 but the 7.0
release cycle hasn't reached the point of dealing with packages quite
yet so I'm holding off on that a bit longer.


git-svn-id: https://svn.freebsd.org/base/stable/6@174728 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f