MFC:	sys/vm/vm_kern.c	1.131
When one tries to allocate memory with the M_WAITOK flag and we are short in
address space in kmem map call vm_lowmem event in a loop and wait a bit for
subsystems to reclaim some memory which in turn will reclaim address space as
well.

Note, this is a work-around.

Reviewed by:	alc
Approved by:	alc
Approved by:	re (gnn)


git-svn-id: https://svn.freebsd.org/base/releng/7.0@175399 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f