MFC:	sys/contrib/opensolaris/uts/common/fs/zfs/zvol.c	1.4
Call zil_commit() (if ZIL is not disabled) after every non-read request
(BIO_WRITE and BIO_FLUSH) as it is done is Solaris. The difference is
that Solaris calls it only for sync requests, but we can't say in GEOM
if the request is sync or async, so we do it for every request.

Approved by:	re (gnn)


git-svn-id: https://svn.freebsd.org/base/releng/7.0@175401 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f