MFC rev 1.19
Add a new function is_default_interface() which determines if this
interface is one with the default route (or there isn't one).  Use it to
decide if we should adjust the default route and /etc/resolv.conf.

Fix the delete of the default route.  The if statement was totally bogus
and the delete only worked due to a typo. [1]

Reported by:    Jordan Coleman <jordan at JordanColeman dot com> [1]


git-svn-id: https://svn.freebsd.org/base/stable/7@178089 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f