* gthr-posix.h (__gthread_active_p): Use the Solaris implementation for FreeBSD as well.
This is the fix for __gthread_active_p() returning false positives
which was committed as rev. 1.1.1.8.2.1 to RELENG_7 but now looped
back to the vendor branch via the GCC repository and relicensed to
be GPLv2 by me.
Thanks go to gerald@ for getting the fix approved upstream and for
committing to the GCC repository.

PR:		119289
Approved by:	core


git-svn-id: https://svn.freebsd.org/base/vendor/gcc/dist@178388 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f