MFC.
The final stage of the big CDDL file move. These files were repo
copied in head to their new location under the 'cddl' directories
of src and src/sys and then added for this branch. The build has
been using the files in their new locations for a few days now.


git-svn-id: https://svn.freebsd.org/base/stable/7@178393 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f