Updated the test to handle multiple sinks.  The sinks cut their transmissions by the number of them running so that they do not overwhelm the source.
Added a simple shell script to kick off sinks on multiple hosts as
well as a source on the host where the shell script is run.  The script
also collects the output of all the sinks and the source into files named
for the host on which the tests are run.  A date is appended to each output
file to make it unique per run.


git-svn-id: https://svn.freebsd.org/base/head@178456 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f