Vendor import of the full userland contrib part of DTrace support from OpenSolaris. This commit resets files to match the versions in the OpenSolaris tree as of 2008/04/10.
The changes in this import from the previous import are the ones that
will subsequently re-applied to take files off the vendor branch. This
is unfortunately necessary because the Solaris developers won't allow
FreeBSD support #ifdefs in their source code because that creates
'dead code' (stuff that they never compile).


git-svn-id: https://svn.freebsd.org/base/vendor-cddl/opensolaris/dist/cddl/contrib/opensolaris@178528 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f