Add the statically defined tracing header.
Note that this implementation differs from the one in OpenSolaris, so
it is BSD licensed and can be included anywhere.

The kernel definitions defined here are dependent on the kernel option
KDTRACE_HOOKS so that macros added to the sources are pre-processed
out completely when the DTrace kernel hooks aren't compiled in.


git-svn-id: https://svn.freebsd.org/base/head@179063 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f