MFC: sys/conf/files.sparc64 1.95; sys/conf/files.sun4v 1.14;      sys/sparc64/isa/isa.c 1.21; sys/sparc64/pci/ofw_pci.h 1.12;      sys/sparc64/pci/ofw_pci_if.m 1.10; sys/sparc64/pci/psycho.c 1.76;      sys/sun4v/sun4v/hv_pci.c 1.5
Remove the MD isa_irq_pending() and the underlying PCI-specific
infrastructure. Its only consumer ever was sio(4) and thus was
unused on sparc64 since removing the last traces of sio(4) in
sparc64 configuration files in favor for uart(4) over three
years ago. If similar functionality is required again it should
be brought back as an MD intr_pending() which works for all
busses by using for example interrupt controller hooks.


git-svn-id: https://svn.freebsd.org/base/stable/7@179446 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f