Update to version 9.3.5. It contains the latest bug fixes, updates to root server addresses, and a fix for the vulnerability mentioned here: http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2008-0122
Users of BIND 9.3.x are strongly encouraged to upgrade to this
version. Also, the 9.3.x branch is now in maintenance-only mode.
Users are encouraged to investigate BIND 9.4.x or perhaps 9.5.x.

http://www.isc.org/index.pl?/sw/bind/versions_and_support.php

This udpate is being done by updating the files directly in this
branch rather than an import + MFC because BIND in HEAD is 9.4.x.


git-svn-id: https://svn.freebsd.org/base/stable/6@179502 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f