MFC rev r179288: - Add locking to all filesystem operations in fdescfs and flag it as MPSAFE. - Use proper synhronization primitives to protect the internal fdesc node cache   used in fdescfs. - Properly initialize and uninitalize hash. - Remove unused functions.
Since fdescfs might recurse on itself, adding proper locking to it needed some
tricky workarounds in some parts to make it work. For instance, a descriptor in
fdescfs could refer to an open descriptor to itself, thus forcing the thread to
recurse on vnode locks. Because of this, other race conditions also had to be
fixed.

Approved by:	kib (mentor)
Reviewed by:	kib (mentor)
Tested by:	pho


git-svn-id: https://svn.freebsd.org/base/stable/7@180121 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f