There actually were bugs in the original handling that I missed last night.
Free the children after each pci bus that is searched.  Otherwise we
leak them.  With free in the new place, we also have to free children
before going to done when we find the device we're looking for.

Also, if we can't get the children of a device, just ignore that bus.


git-svn-id: https://svn.freebsd.org/base/head@182065 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f