Restore 256 pty(4) entries.
As discussed with Robert Watson on the src-committers list, it is safer
to keep at least some pty(4) entries in /etc/ttys, for applications that
roll their own PTY allocation routine and only search for BSD-style
PTY's.

This means we've now just toggled the amount of entries for pts(4) and
pty(4).

Requested by:	rwatson


git-svn-id: https://svn.freebsd.org/base/head@182104 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f