MFC: r182122
There's a race in kmem(4) between checking whether a page is resident
in the kernel and copying it out, causing a panic when faulting on a
nofault entry. Handle this case gracefully by letting the kernel copy
functions return EFAULT instead. As such this change addresses the
same problem as r154721 does for i386.


git-svn-id: https://svn.freebsd.org/base/stable/7@182392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f