Close, but not eliminate, a race condition.  It is one that properly designed drivers would never hit, but was exposed in diving into another problem...
When expanding the devclass array, free the old memory after updating
the pointer to the new memory.  For the following single race case,
this helps:

	allocate new memory
	copy to new memory
	free old memory
<interrupt>				read pointer to freed memory
	update pointer to new memory

Now we do
	allocate new memory
	copy to new memory
	update pointer to new memory
	free old memory

Which closes this problem, but doesn't even begin to address the
multicpu races, which all should be covered by Giant at the moment,
but likely aren't completely.

Note: reviewers were ok with this fix, but suggested the use case
wasn't one we wanted to encourage.

Reviewed by:	jhb, scottl.


git-svn-id: https://svn.freebsd.org/base/head@183750 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f