Package installation is handled by starting off with the list of packages the user selected and then recursively installing their dependencies, finally installing the ones the user selected after the recursion unwinds.  Since users often select "high-level" packages that are on a higher numbered disc for the multi-volume release CDROMS this resulted in excessive disc swapping while installing things like kde, gnome, etc.
Cut down on disc swapping by iterating through the disc volumes one at a
time if we notice the package set is on multiple volumes.  If a package
is on a higher volume don't install it yet, but still "process it" so we
get its dependencies installed.  Because of the way the package sets for
releases get assembled we're guaranteed dependencies will be on the same
volume or lower.

Reviewed by:	jhb
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@183921 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f