MFC of a critical fix for the ESB2 adapters in the em driver, however the fix required a shared code change that involves both em and igb, requiring some compatibility changes throughout.
Also, the igb driver now has header split. This is a hardware
assist that is able to seperate the payload and header into
seperate buffers. Using jumbo frames and 4K mbuf clusters, this
should allow the use of ZERO COPY, but that has yet to be
implemented.

Approved by:re


git-svn-id: https://svn.freebsd.org/base/stable/7@185510 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f