When libthr and rtld start up, there are a number of magic spells cast in order to get the symbol binding state "just so".  This is to allow locking to be activated and not run into recursion problems later.
However, one of the magic bits involves an explicit call to _umtx_op()
to force symbol resolution.  It does a wakeup operation on a fake,
uninitialized (ie: random contents) umtx.  Since libthr isn't active, this
is harmless.  Nothing can match the random wakeup.

However, valgrind finds this and is not amused.  Normally I'd just
write a suppression record for it, but the idea of passing random
args to syscalls (on purpose) just doesn't feel right.


git-svn-id: https://svn.freebsd.org/base/head@185728 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f