MFC of 3 separate commits.
Bug fix from Chelsio which addresses the issue of the device resetting
when it sees only received packets.  In some cases where a device only
recieves data it mistakenly thinks that its transmitting side is broken
and resets the device.

Fix a bug with the ael1006 PHY.  The bug shows up as persistent but incomplete
packet loss, of between 10-30%. The fix is to put the PHY into
and take it out of local loopback mode when resetting the interface.

Bug fix to support N310 version of Chelsio cards (board ID 1088).

Obtained from:	Chelsio Inc.
Approved by: re@


git-svn-id: https://svn.freebsd.org/base/stable/7@185805 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f