MFC: 185746 186789 187143
misc fixes and cleanup of the Forth scripts used by the loader,
see the original commits for details.
The only visible change is that now setting

	loader_conf_files="foo bar ${some.var}"

works as expected, and expands variables correctly.
Also, ${variables} can be used consistently in all assignments
in *.conf files managed by the loader.

NOTE: this commit has nothing to do with the pxeboot and /boot/loader
problems that people are seeing on 6.4 and 7.2 -- those are related
to issues with the loader code, not with the scripts that the
loader processes.


git-svn-id: https://svn.freebsd.org/base/stable/7@187222 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f