MFC 175562 from /head
Repeat iostat header after rows-3 instead of a hardcoded 20.

Use ioctl() to get the window size in iostat(8), and force a new
header to be prepended to the output every time the current
window size changes.  Change the number of lines before each
header to `rows - 3' when the terminal is resized, so that the
full terminal length can be used for output lines.


git-svn-id: https://svn.freebsd.org/base/stable/7@187265 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f