176561 | keramida | 2008-02-25 21:06:43 +0200 (Mon, 25 Feb 2008) | 11 lines
Implement from scratch a -l option for du(1), to match the same option
of the GNU utility.  The default behavior of our original `du' is to
count hardlinked files only once for each invocation of the utility.
With the new -l option they count towards the final size every time
they are found.


git-svn-id: https://svn.freebsd.org/base/stable/7@187272 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f