MFC r182364: Introduce the VV_FORCEINSMQ vnode flag. It instructs the insmnque() function to ignore the unmounting and forces insertion of the vnode into the mount vnode list.
Change insmntque() to fail when forced unmount is in progress and
VV_FORCEINSMQ is not specified.

Add an assertion to the insmntque(), requiring the vnode to be
exclusively locked for mp-safe filesystems.
MFC note: because zfs in stable/7 does not follow the insmntque() protocol,
the assertion is commented out.

Use the VV_FORCEINSMQ for the creation of the syncvnode.


git-svn-id: https://svn.freebsd.org/base/stable/7@188610 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f