Merge 189499,189557,189558
This is the radeon r600+ code with fixes.  A few days early,
but it seems to fix some other vblank related issues as well.

189499:
Import support for ATI Radeon R600 and R700 series chips.

Tested on an HD3850 (RV670) on loan from Warren Block.

Currently, you need one of the following for this to be useful:

        x11-drivers/xf86-video-radeonhd-devel (not tested)
        xf86-video-ati from git (EXA works, xv is too fast)
        xf86-video-radeonhd from git (EXA works, xv works)

There is no 3d support available from dri just yet.

189557:
Call the right function for the right chipset.

189558:
        -Make the PCI(E)/AGP calculations consistent
        -Calculate the scratch address correctly


git-svn-id: https://svn.freebsd.org/base/stable/7@189855 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f