While we currently still seem to have a gateway address, for future stability, check for gw to be set before adding the flag and the address to the routing message.
r186308, backed out in r191305, already tried to do that, and in addition
ignore AF_LINK types of gateway addresses to work around a problem that
r167797 had introduced on the kernel side always setting RTF_GATEWAY if a
gateway address was passed into the kernel.
The proper solution for this is still under discussion so I am hesitant to
re-add the special AF_LINK treatment for now.

MFC after:	3 days


git-svn-id: https://svn.freebsd.org/base/head@191306 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f