MFC r190324:
Support for a new environment variable, LD_ELF_HINTS_PATH for overriding
the rtld hints file.  This environment variable would be unset if the
process is considered as tainted with setuid/setgid.  This feature gives
a convenient way of using a custom set of shared library that is not
located in the default location and switch back.

Feature requested by:	iXsystems
Original patch by:	John Hixson
MFC after:		2 weeks


git-svn-id: https://svn.freebsd.org/base/stable/7@192008 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f