MFC	196234
In function ip_output(), the cached route is flushed when there is a
mismatch between the cached entry and the intended destination. The
cached rtentry{} is flushed but the associated llentry{} is not. This
causes the wrong destination MAC address being used in the output
packets. The fix is to flush the llentry{} when rtentry{} is cleared.

Reviewed by:	kmacy, rwatson
Approved by:	re


git-svn-id: https://svn.freebsd.org/base/stable/8@196235 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f