Add support for backing up the old kernel when installing a new kernel using freebsd-update.  This applies to using freebsd-update in "upgrade mode" and normal freebsd-update on a security branch.
The backup kernel will be written to /boot/kernel.old, if the directory
does not exist, or the directory was created by freebsd-update in a
previous backup.  Otherwise freebsd-update will generate a new directory
name for use by the backup.  By default symbol files are not backed up
to save diskspace and avoid filling up the root partition.

This feature is fully configurable in the freebsd-update config file,
but defaults to enabled.

MFC after:	1 week (stable/7)
Reviewed by:	cperciva
Approved by:	re (kib)


git-svn-id: https://svn.freebsd.org/base/head@196392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f