The boot loader is a FreeBSD a.out binary for x86, not a VAX binary. Rather than writing out a MID of '0', write a MID of 0x86 (aka MID_I386) so that file gets it right.
This is a nop for boot2.  It just checks the MAGIC part of the field,
ignoring the MID.  boot2 is the only thing that loads this file, and
only on x86 so the MID_i386 is always the right value (the rest of the
code is already x86 specific).

Reviewed by:	bde@, jhb@
MFC after:	8.0 is out the door :)


git-svn-id: https://svn.freebsd.org/base/head@197051 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f