MFC: 194246, 194904, 194973
Add cas(4), a driver for Sun Cassini/Cassini+ and National Semiconductor
DP83065 Saturn Gigabit Ethernet controllers. These are the successors
of the Sun GEM controllers and still have a similar but extended transmit
logic. As such this driver is based on gem(4).
Thanks to marcel@ for providing a Sun Quad GigaSwift Ethernet UTP (QGE)
card which was vital for getting this driver to work on architectures
not using Open Firmware.


git-svn-id: https://svn.freebsd.org/base/stable/7@197343 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f