MFC r193375, r194387, r195003:
Add vm_lowmem event handler for dirhash. This will cause dirhashes to be
deleted when the system is low on memory. This ought to allow an
increase to vfs.ufs.dirhash_maxmem on machines that have lots of memory,
without degrading performance by having too much memory reserved for
dirhash when other things need it. The default value for dirhash_maxmem
is being kept at 2MB for now, though.

Approved by:	dwmalone (mentor)


git-svn-id: https://svn.freebsd.org/base/stable/7@197852 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f