Intel XScale hwpmc(4) support.
This brings hwpmc(4) support for 2nd and 3rd generation XScale cores.
Right now it's enabled by default to make sure we test this a bit.
When the time comes it can be disabled by default.
Tested on Gateworks boards.

A man page is coming.

Obtained from:	//depot/user/rpaulo/xscalepmc/...


git-svn-id: https://svn.freebsd.org/base/head@200928 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f