MFC 201215: Add support for configuring vlan(4) interfaces as child devices similar to wlan(4) interfaces.  vlan(4) interfaces are listed via a new 'vlans_<IF>' variable.  If a vlan interface is a number, then that number is treated as the vlan tag for the interface and the interface will be named '<IF>.<tag>'. Otherwise, the vlan tag must be provided via a vlan parameter in a 'create_args_<vlan>' variable.
While I'm here, fix a few nits in rc.conf(5) and mention create_args_<IF> in
the description of cloned_interfaces.

The changes for 7.x are a bit larger as they introduce childif_create() and
childif_destroy() and move ifn_start() and ifn_stop() from rc.d/netif to
network.subr.


git-svn-id: https://svn.freebsd.org/base/stable/7@202169 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f