MFC 189698,192035: Optimize the inner loop of pmap_copy().
Correct a rare use-after-free error in pmap_copy().  This error was
introduced in amd64 revision 1.540 and i386 revision 1.547.  However, it
had no harmful effects until after a recent change, r189698, on amd64.
(In other words, the error is harmless in RELENG_7.)

The error is triggered by the failure to allocate a pv entry for the one
and only mapping in a page table page.  I am addressing the error by
changing pmap_copy() to abort if either pv entry allocation or page
table page allocation fails.  This is appropriate because the creation of
mappings by pmap_copy() is optional.  They are a (possible) optimization,
and not a requirement.

Correct a nearby whitespace error in the i386 pmap_copy().


git-svn-id: https://svn.freebsd.org/base/stable/7@202837 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f