Fix a bug in previous revision.
The bc(1) program may need to deal with files when it's being run in
interactive mode, so we can not blindly use interactive mode (in turn
use libedit) but need to check if the input source is really the standard
input.

This commit should fix a regression where 'bc -l' would not parse the
mathlib.

Reported by:	trasz


git-svn-id: https://svn.freebsd.org/base/head@203531 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f