MFC	205077
The flow-table module retrieves the destination and source
address as well as the transport protocol port information
from the outbound packets. The routing code is generic and
compares every byte in the given sockaddr object. Therefore
the temporary sockaddr objects must be cleared due to padding
bytes. In addition, the port information must be stripped
or the route search will either fail or return the incorrect
route entry.

Unit testing is done using OpenVPN over the if_tun interface.


git-svn-id: https://svn.freebsd.org/base/stable/8@206055 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f