This is an MFC of 205847, 204071, 196580 and 196579
------------------------------------------------------------------------
Change how multipath labels are created and managed. This makes it easier
to support various storage boxes which really aren't active-active.

We only write the label on the *first* provider. For all other providers
we just "add" the disk. This also allows for an "add" verb.

A usage implication is that you should specificy the currently active
storage path as the first provider.

Note that this does not add RDAC-like functionality, but better allows for
autovolumefailover configurations (additional checkins elsewhere will support
this).
------------------------------------------------------------------------

- Style fixes.
- Prefer strlcpy() over strncpy().

------------------------------------------------------------------------

There's no need for checking result of M_WAITOK allocation.

------------------------------------------------------------------------

Fix an obvious topology lock leak.


git-svn-id: https://svn.freebsd.org/base/stable/7@207133 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f