MFC r206773: ln: Do not delete a file by hardlinking it to itself.
Two pathnames refer to the same directory entry iff the directories match
and the final components' names match.

Example: (assuming file1 is an existing file)
  ln -f file1 file1
This now fails while leaving file1 intact. It used to delete file1 and then
complain it cannot be linked because it is gone.

With -i, this error is detected before the question is asked.


git-svn-id: https://svn.freebsd.org/base/stable/8@207464 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f