rc_quiet does not exist in RELENG_7, so hand-merge the part of r179870 that I did not think was applicable here (namely bailing out of run_rc_command before printing the warning) but use rc_fast as the trigger instead of rc_quiet. The former exists in all supported branches, and is set at boot time so it does the job.
The only remaining element that isn't already covered by 'faststart' is
an instance of '/etc/rc.d/ipfilter resync' in rc.d/netif, so make that
fastresync to also silence the warning there. This is similar to what
was done in r175676, although that used quietresync (as described above).

Pointed out by:	jhell <jhell@dataix.net>


git-svn-id: https://svn.freebsd.org/base/stable/7@207811 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f