MFC r208994:
Do not require pos parameter to be zero in MAP_ANONYMOUS mmap requests
in Linux emulation layer. Linux seems to only require that pos is
page-aligned, but otherwise ignores it. Default FreeBSD mmap parameter
checking is too strict to allow some Linux binaries to run. tsMuxeR is
one example of such a binary.

Discussed with:	jhb


git-svn-id: https://svn.freebsd.org/base/stable/8@209304 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f