Simple compatibility hacks for building on older systems where MACHINE_CPUARCH isn't defined.  I believe that this will cover all options.
I didn't define it in kern.mk because $M is set to MACHINE_CPUARCH and
then is expanded for the genassym.o rule in kern.post.mk and kern.mk
is included after this, so the expansion isn't quite right.  I think
this is a bug in make, but don't have the time to track it to ground
(and even if I did, fixing it would require a MFC of the change to the
very old systems we're targetting with this fix).


git-svn-id: https://svn.freebsd.org/base/head@210151 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f