MFC r209738:
Mention the radiolist option in the man page. It is being listed when
dialog(1) is run without arguments and works as expected. Therefore,
it should be part of the manual as well.

Note: dialog(1) has not been updated for many years and is not actively
maintained at the moment.

PR:             docs/139682
Submitted by:   manolis@
Discussed with: jkim@


git-svn-id: https://svn.freebsd.org/base/stable/8@210297 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f