MFC: 210621
Make sure that we clear the correct bits when we turn off
a PMC.  It was possible that we could have turned a bit on but
never cleared it.

Extend the calls to rdmsr() to all necessary functions, not
just those which previously caused a panic.

Pointed out by: jhb@


git-svn-id: https://svn.freebsd.org/base/stable/8@210893 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f