MFC: r173852
Add the -W options, which acts the same as -w but will generate
unique names based on the submitted filename, a strftime(3) format
string and a two digit sequence number.

By default the strftime(3) format string is %Y%m%d (YYYYMMDD), but
this can be changed by the -F option.

PR:		bin/106049 (based on patch in that PR)
Approved by:	grog@ (mentor)


git-svn-id: https://svn.freebsd.org/base/stable/7@213027 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f