/etc/rc.d/defaultroute currently bails immediately if all interfaces set to use DHCP have no carrier.  This can cause grief as it may take some time for link to be established, and defaultroute may terminate before this happens.
Introduce a defaultroute_carrier_delay variable and then wait that long
in defaultroute before bailing if no interfaces have carrier.  With the
default settings defaultroute will wait for five seconds for this, and
the original 30 second wait for a default route to appear is unchanged.
Note that there is in discussion an alternative approach to the broader
problem of waiting for DHCP-configured routes.  However, this change
addresses a real problem in the current defaultroute script.

Discussed on:	freebsd-rc@


git-svn-id: https://svn.freebsd.org/base/head@213270 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f