Split the root mount logic from the (generic) mount code and move it (the root mount code) into a new file called vfs_mountroot.c
The split is almost trivial, as the code is almost perfectly
non-intertwined. The only adjustment needed was to move the UMA
zone allocation out of vfs_mountroot() [in vfs_mountroot.c] and
into vfs_mount.c, where it had to be done as a SYSINIT [see
vfs_mount_init()].

There are no functional changes with this commit.


git-svn-id: https://svn.freebsd.org/base/head@213365 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f