MFC r200288, r200346, r200352:  Make minidump work on i386/XEN.
Note that r200288 and r200346 touched i386/i386/dump_machdep.c,
but in that file r200346 was a back-out of r200288; so the fact
that dump_machdep.c has moved to the x86 tree is not a problem.

In the !XEN case the files under i386/i386 are unchanged after
preprocessing.


git-svn-id: https://svn.freebsd.org/base/stable/8@217050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f