MFC r188498 (partial):
"I've changed main() to set file properly, so output() is never called
with file set to NULL."

Obtained from:	ed

The other bits of that revision are related to using fdevname(3).
This was never merged back to RELENG_7, however the cleanup of
the code in this MFC is good both for general purposes, and diff
reduction to HEAD (which in turn will make the following MFCs easier).


git-svn-id: https://svn.freebsd.org/base/stable/7@217299 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f