MFC r216806: sh: Properly restore exception handler in fc.
If SIGINT arrived at exactly the right moment (unlikely), an exception
handler in a no longer active stack frame would be called.

Because the old handler was not used in the normal path, clang thought it
was a dead value and if an exception happened it would longjmp() to garbage.
This caused builtins/fc1.0 to fail if histedit.c was compiled with clang.
(Note: not tested on stable/8 with clang.)


git-svn-id: https://svn.freebsd.org/base/stable/8@217485 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f