Add a utility, utxrm(8).
Most of the ports I broke when I imported utmpx, were simple management
utilities for the utmp database, allowing you to add/remove entries
manually.

Add a small tool called utxrm(8), which allows you to remove an entry
from the utmpx database by hand. This is useful when a login daemon
crashes or fails to remove the entry during shutdown.


git-svn-id: https://svn.freebsd.org/base/head@218847 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f