MFC r203164:
Add "maxfilesize" mount option for tmpfs to allow specifying the
maximum file size limit. Default is UINT64_MAX when the option is
not specified. It was useless to set the limit to the total amount of
memory and swap in the system.

Use tmpfs_mem_info() rather than get_swpgtotal() in tmpfs_mount() to
check if there is enough memory available.

Remove now unused get_swpgtotal().

MFC r203169:

Bump .Dd for r203164.


git-svn-id: https://svn.freebsd.org/base/stable/8@219429 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f