MFC r209710:
Extend the kernel unit number allocator for allocating specific unit
numbers. This change adds a new function alloc_unr_specific() which
returns the requested unit number if it is free. If the number is
already allocated or out of the range, -1 is returned.

Update alloc_unr(9) manual page accordingly and add a MLINK for
alloc_unr_specific(9).


git-svn-id: https://svn.freebsd.org/base/stable/8@219642 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f