Make freebsd-update(8) smarter in how it handles $FreeBSD$ tags in configuration files.
If the current file differs from the canonical version from the old release
only due to differences in the $FreeBSD$ tag (which can happen if the system
was installed from source code, depending on how the src tree was checked out)
then freebsd-update will treat the file as "unmodified" and silently update
it to the "clean" version in the new release.

If the only change being made to a configuration file is in the $FreeBSD$
tag (e.g., for any configuration files which have been modified locally, now
that we're using SVN and the $FreeBSD$ tag changes when a branch is created),
freebsd-update will no longer print the diff and prompt "Does this look
reasonable (y/n)?".

Nagged by:	pgollucci
MFC after:	1 month


git-svn-id: https://svn.freebsd.org/base/head@221780 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f