sh: Reduce unnecessary forks with eval.
The eval special builtin now runs the code with EV_EXIT if it was run
with EV_EXIT itself.

In particular, this eliminates one fork when a command substitution contains
an eval command that ends with an external program or a subshell.

This is similar to what r220978 did for functions.


git-svn-id: https://svn.freebsd.org/base/head@223163 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f