libprocstat: For MAP_PRIVATE, do not consider the file open for writing.
If a file is mapped with with MAP_PRIVATE, no write permission is required
and changes do not end up in the file. Therefore, tools like fuser and fstat
should not show the file as open for writing.

The protection as displayed by procstat -v still includes write in this
case, and shows 'C' for copy-on-write.


git-svn-id: https://svn.freebsd.org/base/head@223279 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f