MFC: r218745
Remove reading of symbols from a.out loaded files.  Since we are tight
on space for clang and a.out support is only needed for /boot/loader,
they are excess bytes that serve no useful purpose other than to
support really old kernels (FreeBSD < 3.2 or so).  Prefer clang
support over support for these old kernels and remove this code.  We
gain about 100 bytes of space this way.

Reviewed by:	rdivacky@


git-svn-id: https://svn.freebsd.org/base/stable/7@224433 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f