MFC r213736: generic_stop_cpus: prevent parallel execution
Addendum to the original commit message:
This is mostly useful to protect kdb_trap() from concurrent entry as it
doesn't have any protection like e.g. panic(9) has.  Even if kdb_trap()
had that protection there would still be a need for protection
between concurrect panic() and kdb_trap() entries.  That protection
could be implemented externally, but doing it in generic_stop_cpus()
seems to be OK as well (stopping CPUs on panic is in the works).


git-svn-id: https://svn.freebsd.org/base/stable/8@225580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f