MFC r202992:
Cache failing and opened catalogs in catopen() and related functions.
Continuous catopen() calls cause 4 failig stat(2) each, which means a lot
of overhead.  It is also a good idea to keep the opened catalogs in the memory
to speed up further catopen() calls to the same catalog since these catalogs
are not big at all.  In this case, we count references and only free() the
allocated space when the reference count reaches 0.  The reads and writes to
the cache are syncronized with an rwlock when these functions are called from
a threaded program.

MFC r202993, r203174, r203719, r204110:

Small fixes and style nits for the above change.

Approved by:	delphij (mentor)


git-svn-id: https://svn.freebsd.org/base/stable/8@225733 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f