MFC 225223
When an interface address route is removed from the system, another
route with the same prefix is searched for as a replacement. The
current code did not bypass routes that have non-operational
interfaces. This patch fixes that bug and will find a replacement
route with an active interface.

PR:		kern/159603
Submitted by:	pluknet, ambrisko at ambrisko dot com
Reviewed by:	discussed on net@


git-svn-id: https://svn.freebsd.org/base/stable/8@226239 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f