MFC 225163
When the RADIX_MPATH kernel option is enabled, the RADIX_MPATH code tries
to find the first route node of an ECMP chain before executing the route
command. If the system has a default route, and the specific route argument
to the command does not exist in the routing table, then the default route
would be reached. The current code does not verify the reached node matches
the given route argument, therefore erroneous removed the entry. This patch
fixes that bug.


git-svn-id: https://svn.freebsd.org/base/stable/8@226241 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f