Part 1 of two parts, this is the shared code changes in support of new deltas for both em and igb drivers.
Note that I am not able to track all the bugs fixed in
this code, I am a consumer of it as a component of my
core drivers. It is important to keep the FreeBSD drivers
up to date with it however.

One important note is there is a key fix for 82574 in this
update. Also, there are lots of white space changes, I am
not happy about them but have no control over it :)


git-svn-id: https://svn.freebsd.org/base/head@228386 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f