Migrate ufs and ext2fs from skpc() to memcchr().
While there, remove a useless check from the code. memcchr() always
returns characters unequal to 0xff in this case, so inosused[i] ^ 0xff
can never be equal to zero. Also, the fact that memcchr() returns a
pointer instead of the number of bytes until the end, makes conversion
to an offset far more easy.


git-svn-id: https://svn.freebsd.org/base/head@229200 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f