MFC r228801:
Remove -mfancy-math from LIB32CPUFLAGS for amd64.  It has been default
for our gcc since more than three years (see r181534, which is also in
stable/9 and stable/8).  This flag used to be for the benefit of the old
in-kernel math emulator, which was removed more than eight years ago.

Pointed out by:	arundel


git-svn-id: https://svn.freebsd.org/base/stable/9@229277 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f