Prepare for the removal of set_rcvar() by changing the rcvar= assignments to the literal values it would have returned.
The concept of set_rcvar() was nice in theory, but the forks
it creates are a drag on the startup process, which is especially
noticeable on slower systems, such as embedded ones.

During the discussion on freebsd-rc@ a preference was expressed for
using ${name}_enable instead of the literal values. However the
code portability concept doesn't really apply since there are so
many other places where the literal name has to be searched for
and replaced. Also, using the literal value is also a tiny bit
faster than dereferencing the variables, and every little bit helps.


git-svn-id: https://svn.freebsd.org/base/head@230099 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f