MFC r217554:
Specify a CTLTYPE_FOO so that a future sysctl(8) change does not need
to rely on the format string.  For SYSCTL_PROC instances that I
noticed a discrepancy between the CTLTYPE and the format specifier,
fix the CTLTYPE.

The commit is from mdf@ and the MFC has been discussed with him.


git-svn-id: https://svn.freebsd.org/base/stable/8@231391 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f