MFC: r232822
Fix a bug introduced in r223938 (MFC'ed to stable/8 in r224411); on
big-endian machines coping a 32-bit quantum bytewise to the address of
a 64-bit variable results in writing to the "wrong" 32-bit half so adjust
the address accordingly. This fix is implemented in a hackish way for
two reasons:
o in order to be able to get it into 8.3 with zero impact on the little-
  endian architectures where this bug has no effect and
o to avoid blowing the x86 boot2 out of the water again when compiling
  it with clang, which all sane versions of this fix tested do.
This change fixes booting from UFS1 file systems on big-endian machines.


git-svn-id: https://svn.freebsd.org/base/stable/8@232963 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f