We seem to be having issues with a bug in the loader that causes it to estimate the program header size for the kernel wrong.  It can result in the kernel build failing with the message:
ld: kernel.debug: Not enough room for program headers (allocated 5, need 6)
ld: final link failed: Bad value

Tinderbox builds were failing with this message a short time ago, and
I encountered it while preparing the i386 build machine for the 8.3-RC2
builds.

Konstantin has looked at it and believes the issue was fixed in recent
versions of binutils.  It is a bit too late in the release process to
be messing around with the loader and/or binutils.  This workaround
seems to take care of the problem for now.

Reviewed by:	kib


git-svn-id: https://svn.freebsd.org/base/stable/8@233211 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f