MFC r232745:
Add casts to __uint16_t to the __bswap16() macros on all arches which
didn't already have them.  This is because the ternary expression will
return int, due to the Usual Arithmetic Conversions.  Such casts are not
needed for the 32 and 64 bit variants.

While here, add additional parentheses around the x86 variant, to
protect against unintended consequences.


git-svn-id: https://svn.freebsd.org/base/stable/9@233349 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f