CFI fixes for big endian archs.
The flash commands and responses are little-endian and have to be
byte swapped on big-endian systems.  However the raw read of data
need not be swapped.

Make the cfi_read and cfi_write do the swapping, and provide a
cfi_read_raw which does not byte swap for reading data from
flash.


git-svn-id: https://svn.freebsd.org/base/head@233553 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f