Fix a bug in remquo{,f,l}, in which the quotient didn't always have the correct sign when the remainder was 0.
Fix a separate bug in remquo alone, in which the remainder and
quotient were both off by a bit in certain cases involving subnormal
remainders.

The bugs affected all platforms except amd64 and i386, on which the
routines are implemented in assembly.

PR:		166463
Submitted by:	Ilya Burylov
MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@233973 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f