zfsboot: honor -q if it's present in boot.config
Before r228267 the option was honored but the original content of
boot.config was not preserved.  I tried to fix that but missed the idea.
Now the proper way of doing things is taken from i386/boo2.
Also, a comment is added to explain this a little bit unobvious
behavior.

Inspired by:	jhb
MFC after:	5 days


git-svn-id: https://svn.freebsd.org/base/head@234339 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f