Partial merge of r213507 from head:
In r178993, the ability to view a man page by giving a path to the
man page itself was added, but never documented.  Subsequently, the
GNU-licensed man(1) and its man page were replaced by BSD licensed
implementations, and later on this undocumented functionality was
also added to the new BSD-licensed man(1), together with
documentation.

This is a merge of the documentation part of that commit (r213507)
into the GNU-licensed man(1) man page in stable/7, in order to
document the available functionality.

PR:		docs/158973
Merge help:	bz
Approved by:	gabor (mentor)


git-svn-id: https://svn.freebsd.org/base/stable/7@236907 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f