MFC r236578:
Fix build of aicasm when CC=clang.  This was due to a side-effect of the
EARLY_BUILD macro: the -Qunused-arguments flag isn't passed anymore when
building this particular program.  However, with clang 3.1 and -Werror,
such unused argument warnings are flagged as errors, causing buildkernel
to fail at this stage, due to the -nostdinc flag passed during linking.
Since the -nostdinc flag isn't actually needed, just remove it.


git-svn-id: https://svn.freebsd.org/base/stable/9@236975 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f