Do not requeue held page or page for which locking failed, just leave them alone.
Process the act_count updates for the held pages in the vm_pageout
loop over the inactive queue, instead of refusing to do anything with
such page.

Clarify the intent of the addl_page_shortage counter and change its
use for pages which are not processed in the loop according to the
description.

Reviewed by:	alc
MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@238791 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f