MFC r239095: Do not add handler to event handlers list until ithread is created.
In rare event when fast and ithread interrupts share the same vector
and the fast handler was registered first, we can end up trying to
schedule the ithread that is not created yet. The kernel built with
INVARIANTS then triggers an assertion.

Change the order to create the ithread first and only then add the
handler that needs it to the interrupt event handlers list.

Approved by: re (kib)
Reviewed by: jhb


git-svn-id: https://svn.freebsd.org/base/releng/9.1@239437 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f