Fix the PHY / CRC error bug in the AR5212 HAL, which apparently also pops up on (at least) the AR5413.
The 30 second summary - if a CRC error frame comes in during PHY error
processing, that CRC bit will be set for all subsequent frames until
a non-CRC error frame is processed.

So to allow for accurate PHY error processing (Radar, and ANI on the AR5212
HAL chips) just tag the frame as being both CRC and PHY - let the driver
decide what to do with it.

PR:		kern/169362


git-svn-id: https://svn.freebsd.org/base/head@239966 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f