MFC r239178, r239179, r239180, r239181, r239182 and r239299:
Fix detach of USB serial devices so that it doesn't block
the USB explore thread forever. To accomplish this two new
functions have been added to the kernel, to claim and free
the newbus allocated softc.

This change is backwards API compatible, but not binary
compatible. The FreeBSD version has been bumped to force
a recompile of all kernel modules.

Discussed with:         kib, ed, jhb


git-svn-id: https://svn.freebsd.org/base/stable/9@240659 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f