MFC r241401, r241416.
r241401:
Add an "i2c" subcommand to cxgbetool.  You can use this to read
information from the transceivers connected to the ports of a
cxgbe(4) based card.

# cxgbetool t4nex0 i2c <port_id> <dev_addr> <addr> [<count>]

For example:
# cxgbetool t4nex0 i2c 0 0xa0 3
0x10 [16]

(As per SFF-8472 the SFP+ module is at 0xa0 and bit 4 in the value at
address 3 indicates it's a 10Gbase-SR module, which it is.)

r241416:
Add a "clearstats" subcommand to cxgbetool that lets you clear the
MAC statistics for any port.

For example:
# cxgbetool t4nex0 clearstats 0


git-svn-id: https://svn.freebsd.org/base/stable/9@241574 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f