zfs_mount: drop vfs.zfs.rootpool.prefer_cached_config tunable
It turned out to be not that useful, because its default value may lead
to a problem when a root pool is present in zpool.cache, but its
on-disk status is 'exported'.  This may happen if the pool was imported
in a different environment with -f flag and then exported.

MFC after:	12 days


git-svn-id: https://svn.freebsd.org/base/head@242567 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f