Add CTLFLAG_STATS to sysctl flags
In preparation for sysctl(8) growing the ability to only print
out boot/run-time tunables we need a way to differentiate between
RW sysctl nodes that tune a particular thing, or simply export
a stat that we want to allow the sysadmin to reset to 0 (or some
other value).

To do so, we add the CTLFLAG_STATS which should be OR'd into the
CTLFLAGs when exporting a "writable/resettable" statistic node via
sysctl.


git-svn-id: https://svn.freebsd.org/base/head@244101 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f