MFC r244562,245241,245435
Add NetBSD's mtree to the tree and install it as nmtree.

Always install our mtree as /usr/sbin/fmtree and link it as
/usr/sbin/mtree by default.

Add a src.conf option WITH_NMTREE that causes NetBSD's mtree to be linked
as /usr/sbin/mtree as well as /usr/sbin/nmtree.

Sponsored by:	DARPA, AFRL


git-svn-id: https://svn.freebsd.org/base/stable/9@245462 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f