uart: add resume method and enable it for attachments on the most common x86 buses
Otherwise the uart hardware could be in such a state after the resume
where IER is cleared and thus no interrupts are generated.

This behavior is observed and tested with QEMU, so I am comitting this
change to help with my debugging.
There has been no feedback from users of serial ports on real hardware.

MFC after:	20 days


git-svn-id: https://svn.freebsd.org/base/head@246243 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f