Make ctfconvert work correctly on clang-compiled object files.  Clang puts the full original source filename in the STT_FILE entry of the ELF symbol table, while gcc saves only the basename.
Since the DWARF DW_AT_name attribute contains the full source filename,
both for clang and gcc, ctfconvert takes just the basename of it, for
matching with the STT_FILE entry.  So when attempting to match with such
an entry, use its basename, if necessary.

Reported by:	avg
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@247960 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f