Commit the removal of a whitespace to record the proper commit message for the r248519:
For the cam-attached HBAs, allow the driver to specify that it accepts
the unmapped bio by the PIM_UNMAPPED flag.  The CAM passes the
CAM_DATA_BIO data transfer type request for the unmapped bio, and the
driver could use the bus_dmamap_load_ccb() as a helper to
transparently handle the ccb.

Sponsored by:	The FreeBSD Foundation
Reviewed by:	scottl
Tested by:	pho, scottl


git-svn-id: https://svn.freebsd.org/base/head@248520 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f