Fix N32/N64 register saving by ensuring that all registers resolve to unique values.
There's some confusion about what the n32 assembler API really is
(since on page 9 of the spec they say that t0-t3 don't exist, then
turn around on page 22 and say that t4-t7 don't exist), and this
doesn't touch that.

NetBSD's version of this file follows the convention I used here, and
is likely to be correct.

This should fix gdb/ptrace.


git-svn-id: https://svn.freebsd.org/base/head@249523 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f