Make sure the holding descriptor and link pointer are both freed during a non-loss reset.
When the drain functions are called, the holding descriptor and link pointers
are NULLed out.

But when the processq function is called during a non-loss reset, this
doesn't occur.  So the next time a DMA occurs, it's chained to a descriptor
that no longer exists and the hardware gets angry.

Tested:

* AR5416, STA mode; use sysctl dev.ath.X.forcebstuck=1 to force a non-loss
  reset.

TODO:

* Further AR9380 testing just to check that the behaviour for the EDMA
  chips is sane.

PR:		kern/178477


git-svn-id: https://svn.freebsd.org/base/head@250444 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f