MFC r250327
Add a sysctl vfs.read_min to complement the exiting vfs.read_max.  It
defaults to 1, meaning that it's off.

When read-ahead is enabled on a file, the vfs cluster code deliberately
breaks a read into 2 I/O transactions; one to satisfy the actual read,
and one to perform read-ahead.  This makes sense in low-latency
circumstances, but often produces unbalanced i/o transactions that
penalize disks.  By setting vfs.read_min, we can tell the algorithm to
fetch a larger transaction that what we asked for, achieving the same
effect as the read-ahead but without the doubled, unbalanced transaction
and the slightly lower latency.  This significantly helps our workloads
with video streaming.

Submitted by:   emax
Reviewed by:    kib
Obtained from:  Netflix


git-svn-id: https://svn.freebsd.org/base/stable/9@250906 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f