MFC	r251809, r251952, r252012:
Re-sort ext2fs headers to make things easier to find.

In the ext2fs driver we have a mixture of headers:

- The ext2_ prefixed headers have strong influence from NetBSD
and are carry specific ext2/3/4 information.
- The unprefixed headers are inspired on UFS and carry implementation
specific information.

Do some small adjustments so that the information is easier to
find coming from either UFS or the NetBSD implementation.

Also
- Set MAXMNTLEN nearer to where it is used.
- Move EXT2_LINK_MAX to ext2_dir.h .
- Rename some prefixes in the Block Group Descriptor fields.


git-svn-id: https://svn.freebsd.org/base/stable/9@252099 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f