Fix the build with gcc.
Gcc outputs pre-UAL asm and expects the ldcl instruction with a condition
in the form ldc<c>l, where the code produces the instruction in the UAL
form ldcl<c>. Work around this by checking if we are using clang or gcc and
adjusting the instruction.

While here correct the cmp instruction's value to include the # before the
immediate value.


git-svn-id: https://svn.freebsd.org/base/head@252837 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f