There are several code sequences like       vfs_busy(mp);       vfs_write_suspend(mp); which are problematic if other thread starts unmount between two calls.  The unmount starts a write, while vfs_write_suspend() drain writers.  On the other hand, unmount drains busy references, causing the deadlock.
Add a flag argument to vfs_write_suspend and require the callers of it
to specify VS_SKIP_UNMOUNT flag, when the call is performed not in the
mount path, i.e. the covered vnode is not locked.  The suspension is
not attempted if VS_SKIP_UNMOUNT is specified and unmount is in
progress.

Reported and tested by:	Andreas Longwitz <longwitz@incore.de>
Sponsored by:	The FreeBSD Foundation
MFC after:	3 weeks


git-svn-id: https://svn.freebsd.org/base/head@253106 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f