MFC: r253284, r253285, r253303:
Correct the Intel network driver module builds. They were not
defining INET or INET6, and in the case of ixgbe this will cause
a panic in the TSO setup code, but in all cases the ioctl behavior
is different, this change makes the module and static consistent.

Approved by: re


git-svn-id: https://svn.freebsd.org/base/stable/9@253374 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f