In pci_cfgregread() and pci_cfgregwrite(), multiplex the domain and bus number into the bus argument. The bus number occupies the least significant 8 bits. The PCI domain occupies the most significant 24 bits.
On the Altix 350, the PCI domain is a required parameter, but
changing the prototype of the pci_cfgreg*() functions to include a
separate domain argument has wide-spread consequences across the
supported architectures. We'd be changing a known interface.

Multiplexing is an acceptable kluge to give us what we need with
manageable impact. Note that the PCI bus number fits in 8 bits,
so the multiplexing of the domain is a backward compatible change.


git-svn-id: https://svn.freebsd.org/base/head@253560 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f