MFC	r252435, r252437, r253163: Change i_gen in UFS to an unsigned type.
In UFS, i_gen is a random generated value and there is not way for
it to be negative. Actually, the value of i_gen is just used to
match bit patterns and it is of no consequence if the values are
signed or not. Following other filesystems, set it to unsigned,

Calculation for older filesystems remains untouched.


git-svn-id: https://svn.freebsd.org/base/stable/9@254152 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f