Make dummynet use new direct callout(9) execution mechanism.  Since the only thing done by the dummynet handler is taskqueue_enqueue() call, it doesn't need extra switch to the clock SWI context.
On idle system this change in half reduces number of active CPU cycles and
wakes up only one CPU from sleep instead of two.

I was going to make this change much earlier as part of calloutng project,
but waited for better solution with skipping idle ticks to be implemented.
Unfortunately with 10.0 release coming it is better get at least this.


git-svn-id: https://svn.freebsd.org/base/head@254781 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f