Convert the if_lagg rwlock to an rmlock.
We've been seeing lots of cache line contention (but not lock contention!)
in our workloads between the various TX and RX threads going on.

The write lock is only grabbed when configuration changes are made - which
are infrequent.

With this patch, the contention and cycles spent waiting for updates
disappear.

Sponsored by:	Netflix, Inc.


git-svn-id: https://svn.freebsd.org/base/head@255038 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f