Add support for postponing VirtIO virtqueue interrupts
Partial support for the EVENT_IDX feature was added a while ago,
but this commit adds an interface for the device driver to hint
how long (in terms of descriptors) the next interrupt should be
delayed.

The first user of this will be used to reduce VirtIO net's Tx
completion interrupts.


git-svn-id: https://svn.freebsd.org/base/head@255109 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f