Fix for When more than one NIC is present.
The device name was incorrect due to a specific function we ported
from the Linux driver that is not FBSD compatible.  This resulted
with a false sysctl registration and some more problematic issues.

The patch basically revokes it all together.

Submitted by: Meny Yossefi (menyy mellanox.com)

Approved by:	re


git-svn-id: https://svn.freebsd.org/base/head@256269 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f