Fix resource free.
The order of releasing resources in mlxen was wrong, which caused
panic on reload of the module.

conf_ctx list should be released before stat_ctx list, otherwise
the leafs in conf_ctx list won't be released because of the dependancy.

The fix is to change the order of the releases.

Submitted by:	Shahar Klein (shahark at mellanox.com)


git-svn-id: https://svn.freebsd.org/base/head@256682 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f