Merge from projects/arm_eabi_vfp r255380:
Fix the VCVT instruction. It must round towards zero when converting from
a floating-point to an integer value. This was not the case causing issues
when printing certain values.

There is a VCVTR instruction that will round depending on the current
rounding mode. We don't yet support this instruction, or setting the
rounding mode.


git-svn-id: https://svn.freebsd.org/base/head@256790 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f