Provide a working example line for an interface with 1 address running with CARP.
Currently, we've got a problem that interface isn't IFF_UP at the time
we assign it a redundant address, and the latter gets stuck in INIT state.
Additional SIOCSIFFLAGS from ifconfig(8) kicks it to a working state.

A proper fix is kernel side and appeared to be non-trivial, not to be
checked in before 10.0-RELEASE.

Submitted by:	Ole Myhre <ole.myhre dataoppdrag.no>


git-svn-id: https://svn.freebsd.org/base/head@256824 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f