MFC: 256555
Move the resource allocation from the ata_*_probe section to the ata_*_attach
section. This prevents a boot crash on nearly all iMacs and PowerMacs/Books.

The allocation in the probe section was working before because ata_probe was
returning 0 which did not invoke a second DEVICE_PROBE. Now it returns
a BUS_PROBE_DEFAULT which can invoke a second DEVICE_PROBE which results in
a "failed to reserve resource" exit.

PR:	powerpc/182978
Approved by:	re(gjb)


git-svn-id: https://svn.freebsd.org/base/stable/10@256857 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f