Notional MFC of r257057:
Regen documentation post r257449:

Stop conflating WITHOUT_CLANG with WITHOUT_CLANG_IS_CC.  This allows
bootstrapping a copy of clang without building clang for the base system
which is useful for nanobsd and similar setups.  It's still probably
wrong to conflate what is installed as /usr/bin/cc with the selection
of a bootstrap compiler under WITH*_CLANG_IS_CC, but that's for another
day.

Sponsored by:	DARPA/AFRL
Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/stable/10@257450 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f