MFC r257658:
For C++ programs, don't emit any explicit dependency to libstdc++.a or
libc++.a during the early build stages (bootstrap-tools, build-tools,
cross-tools), since it is not possible to know in advance which C++
library is available on the host system.

Instead, just use the bootstrap compiler's built-in default.  This
should eventually make it possible to build stable/9 on head, or on
stable/10, which ship without libstdc++ by default.


git-svn-id: https://svn.freebsd.org/base/stable/9@257812 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f